Setup of todo list and improvement of s=craftersmelter and UI
