+++ src/main/java/com/minecolonies/coremod/network/messages/OpenCraftingGUIMessage.java
@@ -4,24 +4,13 @@ import com.minecolonies.api.colony.IColony;
 import com.minecolonies.api.colony.IColonyManager;
 import com.minecolonies.api.colony.buildings.IBuilding;
 import com.minecolonies.api.colony.permissions.Action;
-import com.minecolonies.api.inventory.container.ContainerCrafting;
-import com.minecolonies.api.inventory.container.ContainerCraftingFurnace;
 import com.minecolonies.api.network.IMessage;
-import com.minecolonies.coremod.colony.buildings.AbstractBuildingSmelterCrafter;
 import com.minecolonies.coremod.colony.buildings.views.AbstractBuildingView;
-import io.netty.buffer.Unpooled;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.inventory.container.INamedContainerProvider;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
 import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.network.NetworkEvent;
-import net.minecraftforge.fml.network.NetworkHooks;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -41,11 +30,6 @@ public class OpenCraftingGUIMessage implements IMessage
     private int colonyId;
 
     /**
-     * Size of the crafting grid.
-     */
-    private int gridSize;
-
-    /**
      * The dimension of the 
      */
     private int dimension;
@@ -62,13 +46,11 @@ public class OpenCraftingGUIMessage implements IMessage
      * Creates an open inventory message for a building.
      *
      * @param building {@link AbstractBuildingView}
-     * @param gridSize the grid size.
      */
-    public OpenCraftingGUIMessage(@NotNull final AbstractBuildingView building, final int gridSize)
+    public OpenCraftingGUIMessage(@NotNull final AbstractBuildingView building)
     {
         super();
         this.buildingId = building.getPosition();
-        this.gridSize = gridSize;
         this.colonyId = building.getColony().getID();
         this.dimension = building.getColony().getDimension();
     }
@@ -76,7 +58,6 @@ public class OpenCraftingGUIMessage implements IMessage
     @Override
     public void fromBytes(@NotNull final PacketBuffer buf)
     {
-        this.gridSize = buf.readInt();
         this.colonyId = buf.readInt();
         this.buildingId = buf.readBlockPos();
         dimension = buf.readInt();
@@ -85,7 +66,6 @@ public class OpenCraftingGUIMessage implements IMessage
     @Override
     public void toBytes(@NotNull final PacketBuffer buf)
     {
-        buf.writeInt(this.gridSize);
         buf.writeInt(this.colonyId);
         buf.writeBlockPos(buildingId);
         buf.writeInt(dimension);
@@ -105,50 +85,8 @@ public class OpenCraftingGUIMessage implements IMessage
         final ServerPlayerEntity player = ctxIn.getSender();
         if (colony != null && checkPermissions(colony, player))
         {
-            final BlockPos pos = buildingId;
-            //todo, which is our inventory?
             final IBuilding building = colony.getBuildingManager().getBuilding(buildingId);
-            if (building instanceof AbstractBuildingSmelterCrafter)
-            {
-                NetworkHooks.openGui(player, new INamedContainerProvider()
-                {
-                    @Override
-                    public ITextComponent getDisplayName()
-                    {
-                        return new StringTextComponent("Furnace Crafting GUI");
-                    }
-
-                    @NotNull
-                    @Override
-                    public Container createMenu(final int id, @NotNull final PlayerInventory inv, @NotNull final PlayerEntity player)
-                    {
-                        final PacketBuffer buffer = new PacketBuffer(Unpooled.buffer());
-                        buffer.writeBlockPos(pos);
-                        return new ContainerCraftingFurnace(id, inv, buffer);
-                    }
-                }, pos);
-            }
-            else
-            {
-                NetworkHooks.openGui(player, new INamedContainerProvider()
-                {
-                    @Override
-                    public ITextComponent getDisplayName()
-                    {
-                        return new StringTextComponent("Crafting GUI");
-                    }
-
-                    @NotNull
-                    @Override
-                    public Container createMenu(final int id, @NotNull final PlayerInventory inv, @NotNull final PlayerEntity player)
-                    {
-                        final PacketBuffer buffer = new PacketBuffer(Unpooled.buffer());
-                        buffer.writeBoolean(gridSize > 2);
-                        buffer.writeBlockPos(pos);
-                        return new ContainerCrafting(id, inv, buffer);
-                    }
-                }, buffer -> new PacketBuffer(buffer.writeBoolean(gridSize > 2)).writeBlockPos(pos));
-            }
+            building.openCraftingContainer(player);
         }
     }
